* src/boehm-gc/darwin_stop_world.c, src/boehm-gc/threadlibs.c, src/boehm-gc/doc/README.changes, src/boehm-gc/doc/README, src/boehm-gc/doc/gcinterface.html, src/boehm-gc/mark.c, src/boehm-gc/include/gc.h, src/boehm-gc/include/gc_cpp.h, src/boehm-gc/include/gc_pthread_redirects.h, src/boehm-gc/include/leak_detector.h, src/boehm-gc/include/gc_config_macros.h, src/boehm-gc/include/private/gc_priv.h, src/boehm-gc/include/private/gcconfig.h, src/boehm-gc/include/private/gc_pmark.h, src/boehm-gc/include/private/gc_locks.h, src/boehm-gc/configure.in, src/boehm-gc/allchblk.c, src/boehm-gc/powerpc_darwin_mach_dep.s, src/boehm-gc/dbg_mlc.c, src/boehm-gc/version.h, src/boehm-gc/pthread_stop_world.c, src/boehm-gc/malloc.c, src/boehm-gc/headers.c, src/boehm-gc/pthread_support.c, src/boehm-gc/Makefile.direct, src/boehm-gc/os_dep.c, src/boehm-gc/Makefile.am, src/boehm-gc/dyn_load.c, src/boehm-gc/alloc.c, src/boehm-gc/misc.c: Updated to upstream version 6.7.

+++ src/boehm-gc/headers.c
@@ -212,7 +212,7 @@ register struct hblk * h;
     result = alloc_hdr();
     SET_HDR(h, result);
 #   ifdef USE_MUNMAP
-	result -> hb_last_reclaimed = GC_gc_no;
+	result -> hb_last_reclaimed = (unsigned short)GC_gc_no;
 #   endif
     return(result);
 }
